Move PackageUtil to commons
